[WEBOPS-2146] Add proxy server to webpack configuration
In order to run a backend server for the API along side the webpack-dev-server
in development, we need to proxy `/api` calls through to another port.

By spinning up an express server for the proxy we can pass calls to `/build`
through to the webpack-dev-server on port `:3001` for serving static assets,
and calls to `/api` through to the backend on port `:3002`.